+++ src/sig_basic.c
@@ -18,8 +18,12 @@ void check_gbbasic(void) {
 
     // ==== SHARED CODE WITH C STARTS HERE ====
 
-    entry = FORMAT_ENTRY(TYPE_ENGINE, "GBBasic", "Alpha3");
-    if (CHECK_PATTERN_AT_ADDR(sig_gbbasic_actor_init_alpha3, sig_gbbasic_actor_init_alpha3_at)) {
+    if (CHECK_PATTERN_AT_ADDR(sig_gbbasic_actor_init, sig_gbbasic_actor_init_alpha3_at)) {
+        entry = FORMAT_ENTRY(TYPE_ENGINE, "GBBasic", "Alpha3");
+        entry_add(entry);
+    }
+    else if (CHECK_PATTERN_AT_ADDR(sig_gbbasic_actor_init, sig_gbbasic_actor_init_alpha4_at)) {
+        entry = FORMAT_ENTRY(TYPE_ENGINE, "GBBasic", "Alpha4");
         entry_add(entry);
     }